+++ sdk/translation/Azure.AI.Translation.Document/CHANGELOG.md
@@ -6,6 +6,8 @@
 - Added optional parameter `categoryId` to the `DocumentTranslationInput.AddTarget`.
 
 ### Breaking Changes
+- Renamed type `StorageInputType` to `StorageInputUriKind`.
+- Renamed property `StorageType` to `StorageUriKind` in `DocumentTranslationInput`.
 
 ### Bugs Fixed